Add version info to the titlebar
