+++ cmake/Findgit_version.cmake
@@ -1,0 +1,8 @@
+include(FetchContent)
+FetchContent_Declare(git_version
+  GIT_REPOSITORY https://github.com/martonmiklos/cmake-git-version-tracking.git
+  GIT_TAG 3b792b2d68f85a1e88210ada8a86fd2aa0ece14e
+)
+FetchContent_MakeAvailable(git_version)
+
+
